Extract rendering and plugin integration modules
