+++ src/plugin-entry.js
@@ -1,4 +1,10 @@
-const { Plugin, PluginSettingTab, Setting, Notice } = require("obsidian");
+const {
+  contextual: LEARNER_CONTEXTUAL_GLOSSES,
+  surface: LEARNER_SURFACE_GLOSSES,
+  construction: LEARNER_CONSTRUCTION_GLOSSES,
+} = require("./runtime-resources/presentation/learner-glosses");
+const createLearnerDisplay = require("./render/learner-display");
+const createCantoSpanPlugin = require("./plugin/canto-span-plugin");
 
 /*
  * Canto Span
@@ -51,23 +57,6 @@ const CANTO_SPAN_RUNTIME_VERSION = "0.5.216";
 // v0.5.161: Distinguishes genuinely impersonal environmental clauses from context-dependent omitted-subject predicates and adds transparent ambient location frames.
 // v0.5.157: Accounts bare relational coverbs such as 對 separately from high-confidence topic-frame linkers, while formal 對於 remains an explicit overt topic-frame source.
 // v0.5.155: Preserves resolved clause/question/VP structure beneath overt-object selection review instead of flattening the grammatical host.
-const BLOCK_CLASS = "canto-study-block canto-span-block";
-const GLOBAL_COLOR_ACCESSIBILITY_CLASS = "canto-global-color-accessibility-enabled";
-const GLOBAL_DIAGNOSTICS_CLASS = "canto-global-diagnostics-enabled";
-const GLOBAL_DIAGNOSTICS_EXPANDED_CLASS = "canto-global-diagnostics-expanded";
-
-const ROLE_COLOR_SETTINGS = require("./runtime-resources/presentation/role-colors");
-const DEFAULT_ROLE_COLORS = Object.fromEntries(
-  ROLE_COLOR_SETTINGS.map((role) => [role.key, role.defaultColor])
-);
-
-const DEFAULT_SETTINGS = {
-  showDiagnostics: true,
-  expandDiagnostics: false,
-  showColorAccessibilityMarkers: false,
-  roleColors: { ...DEFAULT_ROLE_COLORS },
-};
-
 // Runtime construction governance is deliberately minimal. Linguistic status,
 // confidence, sources, speaker records, corpus counts, and promotion eligibility
 // live in grammar/<linguistic-status>/*.md and are validated outside the shipped plugin.
@@ -428,13 +417,6 @@ const CATEGORY_SPAN_TEMPLATES = require("./runtime-resources/grammar/templates/c
 const SLOT_ALIASES = require("./runtime-resources/grammar/slot-aliases");
 const PUNCT_RE = require("./parser/tokenization/punctuation");
 
-function createEl(tag, className, text) {
-  const element = document.createElement(tag);
-  if (className) element.className = className;
-  if (text !== undefined && text !== null) element.textContent = String(text);
-  return element;
-}
-
 const {
   normalizeSurface,
   foldedPinyinFalloutRepair,
@@ -556,34 +538,6 @@ function parserShadowRepairKindLabel(row) {
   if (kind === "folded_lexical_repair") return "folded lexical";
   if (kind === "character_form_normalization") return "character form";
   return String(kind || "shadow repair").replace(/_/g, " ");
-}
-
-function renderParserShadowRepairs(analysis, container) {
-  const rows = parserShadowRepairDisplayRows(analysis);
-  if (!rows.length) return;
-  const box = createEl("div", "canto-normalization-shadow-repairs canto-normalization-shadow-repairs-compact");
-  const label = rows.length > 1 ? `Shadow ×${rows.length}:` : "Shadow:";
-  box.appendChild(createEl("span", "canto-normalization-shadow-label", label));
-  box.setAttribute("title", rows.map((row) => {
-    const kindLabel = parserShadowRepairKindLabel(row);
-    const note = row.note || "Parser-shadow repair applied to parser shadow only.";
-    return `${row.display} — ${kindLabel}. ${note} Raw learner-visible text is not replaced.`;
-  }).join("\n"));
-  rows.forEach((row) => {
-    const item = createEl("span", "canto-normalization-shadow-repair");
-    const kindLabel = parserShadowRepairKindLabel(row);
-    const titleParts = [
-      `${row.display} — ${kindLabel}.`,
-      row.note || "Parser-shadow repair applied to parser shadow only.",
-      "Raw learner-visible text is not replaced."
-    ];
-    item.setAttribute("title", titleParts.join(" "));
-    item.appendChild(createEl("span", "canto-normalization-shadow-raw", row.raw));
-    item.appendChild(createEl("span", "canto-normalization-shadow-arrow", "→"));
-    item.appendChild(createEl("span", "canto-normalization-shadow-normalized", row.normalized));
-    box.appendChild(item);
-  });
-  container.appendChild(box);
 }
 
 function safeClass(value) {
@@ -15135,159 +15089,25 @@ function jyutpingAuditSummary(analysis) {
   };
 }
 
-// Learner-safe hover/title strings. v0.5.51 keeps concise role/syntax metadata
-// at the top of token hovers, while raw slots, internal counts, and template
-// notes stay in diagnostics only. Default hover is now surface + role/syntax +
-// Jyutping + plain learner gloss/context.
 const {
-  contextual: LEARNER_CONTEXTUAL_GLOSSES,
-  surface: LEARNER_SURFACE_GLOSSES,
-  construction: LEARNER_CONSTRUCTION_GLOSSES,
-} = require("./runtime-resources/presentation/learner-glosses");
-
-
-function cleanLearnerNote(note) {
-  let text = String(note || "").trim();
-  if (!text) return "";
-  text = text.split(/[.;]/)[0].trim();
-  if (/\b(parser|diagnostic|template|construction|contextual|affordance|slot|syntax|role|internal|retained|transparent|protected|fallback|registry|tokeni[sz]ation|surface|native-speech|implementation)\b/i.test(text)) return "";
-  return text;
-}
-
-function learnerGlossFromLexicon(surface) {
-  const entry = TOKEN_LEXICON[String(surface || "")] || {};
-  return cleanLearnerNote(entry.note);
-}
-
-function contextualLearnerGlossLinesForToken(node) {
-  const surface = node && node.surface ? String(node.surface) : "";
-  const syntax = node && node.syntax ? String(node.syntax) : "";
-  if (surface === "吓" && /(^|\s)delimitative_aspect(\s|$)/.test(syntax)) {
-    return ["try briefly / do a little", "Marks a light action."];
-  }
-  if (surface === "幾" && /(^|\s)scalar_value_question(\s|$)/.test(syntax)) {
-    return ["how much / to what degree", "Asks about a value on this scale."];
-  }
-  if (surface === "呢" && /(^|\s)discourse_fragment_question(\s|$)/.test(syntax)) {
-    return ["what about…? / and…?", "Returns to a topic or asks about a context-supplied alternative."];
-  }
-  if (surface === "呢" && /(^|\s)sentence_final_question_particle(\s|$)/.test(syntax)) {
-    return ["question particle", "Softly asks for an answer or confirmation."];
-  }
-  if (/(^|\s)have_or_not_experiential_question_marker(\s|$)/.test(syntax)) {
-    return ["ever ... or not?", "Asks whether the experience happened."];
-  }
-  if (/(^|\s)have_or_not_event_question_marker(\s|$)/.test(syntax)) {
-    return ["did ... or not?", "Asks whether the event happened."];
-  }
-  if (surface === "嘅話" && /(^|\s)conditional_marker(\s|$)/.test(syntax)) {
-    return ["if / in the case that", "Introduces a condition for a result."];
-  }
-  if (surface === "用" && /(^|\s)intended_function_use_marker(\s|$)/.test(syntax)) {
-    return ["used for", "Part of 用嚟/用來, linking a thing to its stated use or function."];
-  }
-  if (["嚟", "來"].includes(surface) && /(^|\s)intended_function_lai_linker(\s|$)/.test(syntax)) {
-    return ["for / to", "Links the thing to its stated use or function; it is not a movement verb here."];
-  }
-  return null;
-}
-
-function learnerGlossLinesForToken(node) {
-  const parent = node && node.parent ? node.parent : "";
-  const surface = node && node.surface ? node.surface : "";
-  const traceLearnerGloss = node && node.trace && Array.isArray(node.trace.learner_gloss_lines)
-    ? node.trace.learner_gloss_lines.filter(Boolean)
-    : [];
-  if (traceLearnerGloss.length) return traceLearnerGloss;
-  const syntaxContextual = contextualLearnerGlossLinesForToken(node);
-  if (syntaxContextual) return syntaxContextual;
-  const contextual = LEARNER_CONTEXTUAL_GLOSSES[`${parent}|${surface}`];
-  if (contextual) return contextual;
-  const surfaceGloss = LEARNER_SURFACE_GLOSSES[surface];
-  if (surfaceGloss) return surfaceGloss;
-  const lexiconGloss = cleanLearnerNote(node && node.note) || learnerGlossFromLexicon(surface);
-  if (lexiconGloss) return [lexiconGloss];
-  if (node && node.jyutping) return ["Cantonese word."];
-  return ["Cantonese text."];
-}
-
-function learnerGlossLinesForConstruction(node) {
-  if (!node || node.kind !== "construction") return ["Cantonese phrase."];
-  const surfaceGloss = LEARNER_SURFACE_GLOSSES[node.surface || ""];
-  if (surfaceGloss) return surfaceGloss;
-  const trace = node.trace || {};
-  const traceLearnerGloss = Array.isArray(trace.learner_gloss_lines)
-    ? trace.learner_gloss_lines.filter(Boolean)
-    : [];
-  if (traceLearnerGloss.length) return traceLearnerGloss;
-  if (node.type === "ClauseRelationGraph" && trace.topic_chain_status === "licensed_overt_topic_chain") {
-    return ["topic chain", "An overt topic supplies the understood object of later actions in the linked sequence."];
-  }
-  if (node.type === "ClauseRelationGraph" && trace.topic_chain_status === "antecedent_predicate_compatibility_review") {
-    return ["possible topic chain — needs context", "A topic is present, but it does not clearly fit the omitted object of the later action."];
-  }
-  if (trace.particle_cluster_root) {
-    if (node.type === "PolarQuestionFrame") {
-      return ["biased yes-or-no question", "Asks whether the statement is true while preserving the visible inner particle sequence."];
-    }
-    return ["ordered particle phrase", "Shows several sentence-final particles in their visible inside-to-outside scope order."];
-  }
-  if (node.type === "DiscourseParticleFrame") {
-    const family = String(trace.discourse_particle_family || "");
-    if (family === "epistemic_stance") return ["uncertainty-marked statement", "Marks the statement as probable, possible, or uncertain."];
-    if (family === "evidential_noteworthiness") return ["report or notice-marked statement", "Presents the statement as reported, noteworthy, a reminder, or unexpected."];
-    if (family === "directive_interpersonal_closure") return ["prompt or closure phrase", "Uses a final particle for a directive, suggestion, invitation, or conversational closure."];
-    if (family === "change_state_current_relevance") return ["changed-situation statement", "Marks a new, changed, or currently relevant situation."];
-  }
-  const constructionGlossType = node.compatibility_alias || diagnosticCompatibilityConstructionType(node.type || "");
-  const constructionGloss = LEARNER_CONSTRUCTION_GLOSSES[constructionGlossType]
-    || LEARNER_CONSTRUCTION_GLOSSES[node.type || ""];
-  if (constructionGloss) return constructionGloss;
-  const noteGloss = cleanLearnerNote(node.note || "");
-  if (noteGloss) return [noteGloss];
-  return ["Cantonese phrase."];
-}
-
-function compactLearnerHoverLines(lines) {
-  return [...new Set((lines || [])
-    .map((line) => String(line || "").trim())
-    .filter(Boolean))];
-}
-
-function learnerVisibleSyntax(syntax = "") {
-  return String(syntax || "")
-    .split(/\s+/)
-    .filter(Boolean)
-    .filter((part) => !/^phase\d+_/i.test(part))
-    .join(" ");
-}
-
-function learnerUiHoverTitleForToken(node) {
-  if (!node || node.kind !== "token") return "";
-  const role = node.role || node.label || "";
-  const syntax = learnerVisibleSyntax(node.syntax || "");
-  const displaySurface = node.display_surface || node.surface || "";
-  return compactLearnerHoverLines([
-    displaySurface,
-    role ? `role: ${role}` : "",
-    syntax ? `syntax: ${syntax}` : "",
-    node.jyutping || "",
-    ...learnerGlossLinesForToken(node),
-  ]).join("\n");
-}
-
-function learnerUiHoverTitleForConstruction(node) {
-  if (!node || node.kind !== "construction") return "";
-  return compactLearnerHoverLines([
-    node.display_surface || nodeDisplaySurface(node) || node.surface || "",
-    ...learnerGlossLinesForConstruction(node),
-  ]).join("\n");
-}
-
-function learnerUiHoverTitleForConstructionLayer(node) {
-  if (!node || node.kind !== "construction") return "";
-  return learnerUiHoverTitleForConstruction(node);
-}
+  cleanLearnerNote,
+  learnerGlossFromLexicon,
+  contextualLearnerGlossLinesForToken,
+  learnerGlossLinesForToken,
+  learnerGlossLinesForConstruction,
+  compactLearnerHoverLines,
+  learnerVisibleSyntax,
+  learnerUiHoverTitleForToken,
+  learnerUiHoverTitleForConstruction,
+  learnerUiHoverTitleForConstructionLayer,
+} = createLearnerDisplay({
+  TOKEN_LEXICON,
+  LEARNER_CONTEXTUAL_GLOSSES,
+  LEARNER_SURFACE_GLOSSES,
+  LEARNER_CONSTRUCTION_GLOSSES,
+  diagnosticCompatibilityConstructionType,
+  nodeDisplaySurface,
+});
 
 function learnerUiHoverLeakFlags(title) {
   const text = String(title || "");
@@ -16950,780 +16770,39 @@ function formatNoteMarkdownWithDiagnostics(markdown, notePath = "") {
   return `${header}${noteText}`;
 }
 
-async function copyTextToClipboard(text) {
-  if (typeof navigator !== "undefined" && navigator.clipboard && navigator.clipboard.writeText) {
-    await navigator.clipboard.writeText(text);
-    return;
-  }
-  const textarea = document.createElement("textarea");
-  textarea.value = text;
-  textarea.setAttribute("readonly", "");
-  textarea.style.position = "fixed";
-  textarea.style.left = "-9999px";
-  document.body.appendChild(textarea);
-  textarea.select();
-  document.execCommand("copy");
-  textarea.remove();
-}
-
-function renderCopyButton(parent, textFactory, label, successMessage) {
-  const button = createEl("button", "canto-diagnostic-copy canto-diagnostic-ui-only", label);
-  button.type = "button";
-  button.addEventListener("click", async (event) => {
-    event.preventDefault();
-    event.stopPropagation();
-    try {
-      const text = await Promise.resolve(textFactory());
-      if (!String(text || "").trim()) {
-        new Notice("No Canto Span diagnostics found to copy.");
-        return;
-      }
-      await copyTextToClipboard(text);
-      new Notice(successMessage || "Copied Canto Span diagnostic.");
-    } catch (error) {
-      console.error("Canto Span diagnostic copy failed", error);
-      new Notice("Could not copy diagnostic text.");
-    }
-  });
-  parent.appendChild(button);
-  return button;
-}
-
-function renderActionButton(parent, action, label) {
-  const button = createEl("button", "canto-diagnostic-copy canto-diagnostic-ui-only", label);
-  button.type = "button";
-  button.addEventListener("click", async (event) => {
-    event.preventDefault();
-    event.stopPropagation();
-    try {
-      await action();
-    } catch (error) {
-      console.error("Canto Span diagnostic action failed", error);
-      new Notice("Could not run Canto Span diagnostic action.");
-    }
-  });
-  parent.appendChild(button);
-  return button;
-}
-
-function renderPre(parent, className, value) {
-  const pre = createEl("pre", className);
-  pre.textContent = typeof value === "string" ? value : JSON.stringify(value, null, 2);
-  parent.appendChild(pre);
-}
-
-function renderDiagnosticDetails(analysis, container, expanded = false) {
-  const details = createEl("details", "canto-diagnostic");
-  details.open = Boolean(expanded);
-  const summary = diagnosticSummary(analysis);
-  const jyutping = jyutpingAuditSummary(analysis);
-  const reviewNote = summary.review_count
-    ? ` — ${summary.review_count} review trace(s)`
-    : " — no review traces";
-  const jyutpingNote = jyutping.status === "WARN"
-    ? ` — Jyutping WARN (${jyutping.items_needing_review.length})`
-    : " — Jyutping PASS";
-  details.appendChild(createEl("summary", "canto-diagnostic-summary", `Parse diagnostic${reviewNote}${jyutpingNote}`));
-
-  const body = createEl("div", "canto-diagnostic-body");
-  const toolbar = createEl("div", "canto-diagnostic-toolbar");
-  renderCopyButton(toolbar, () => formatDiagnosticMarkdown(analysis), "Copy diagnostic", "Copied this Canto Span diagnostic.");
-  body.appendChild(toolbar);
-
-  body.appendChild(createEl("div", "canto-diagnostic-caption", "Summary"));
-  renderPre(body, "canto-diagnostic-json", summary);
-
-  const foldedRepairRows = foldedLexicalRepairDisplayRows(analysis);
-  if (foldedRepairRows.length) {
-    body.appendChild(createEl("div", "canto-diagnostic-caption", "Folded pinyin-fallout lexical repairs"));
-    renderPre(body, "canto-diagnostic-json", {
-      status: "PASS",
-      policy: "Repairs are applied to parser shadow only. Raw learner-visible text is preserved.",
-      repairs: foldedRepairRows,
-    });
-  }
-  const reviewSuggestionRows = normalizationReviewSuggestionDisplayRows(analysis);
-  if (reviewSuggestionRows.length) {
-    body.appendChild(createEl("div", "canto-diagnostic-caption", "Review-only normalization suggestions"));
-    renderPre(body, "canto-diagnostic-json", {
-      status: "PASS",
-      policy: "Compatibility lane only. Current v0.5.107 pinyin-fallout repairs should not depend on review-only suggestions.",
-      suggestions: reviewSuggestionRows,
-    });
-  }
-
-  body.appendChild(createEl("div", "canto-diagnostic-caption", "Jyutping audit"));
-  renderPre(body, "canto-diagnostic-json", jyutpingAuditSummary(analysis));
-
-  body.appendChild(createEl("div", "canto-diagnostic-caption", "Learner UI hover audit"));
-  renderPre(body, "canto-diagnostic-json", learnerUiHoverAuditSummary(analysis));
-
-  body.appendChild(createEl("div", "canto-diagnostic-caption", "Learner UI hover rows"));
-  renderPre(body, "canto-diagnostic-json", learnerUiHoverAuditRows(analysis));
-
-  body.appendChild(createEl("div", "canto-diagnostic-caption", "Tokenization before construction wrapping"));
-  renderPre(body, "canto-diagnostic-json", diagnosticTokenRows(analysis));
-
-  body.appendChild(createEl("div", "canto-diagnostic-caption", "Final construction tree"));
-  renderPre(body, "canto-diagnostic-json", diagnosticFinalRows(analysis));
-
-  body.appendChild(createEl("div", "canto-diagnostic-caption", "Legend"));
-  renderPre(body, "canto-diagnostic-json", diagnosticLegend());
-
-  details.appendChild(body);
-  container.appendChild(details);
-}
-
-function parseContextDirectiveLine(line, sourceLineIndex) {
-  const raw = String(line || "");
-  // Strip only leading Unicode layout markers for directive recognition. The raw
-  // line and the Cantonese context value remain unchanged for display/export.
-  const trimmed = raw.trim().replace(/^[\uFEFF\u200B-\u200D\u2060]+/u, "");
-  const prefix = trimmed.slice(0, 8);
-  if (prefix.normalize("NFKC").toLowerCase() !== "@context") return null;
-
-  let remainder = trimmed.slice(8).replace(/^[\s\u00A0\u3000]+/u, "");
-  const separator = remainder.charAt(0);
-  if (separator !== ":" && separator !== "：") return null;
-  remainder = remainder.slice(1);
-
-  const contextSource = remainder.trim();
-  return {
-    id: `block-context-${sourceLineIndex}`,
-    directive_kind: "explicit_context_next_source",
-    directive_line_index: sourceLineIndex,
-    raw_directive: raw,
-    separator_code_point: separator.codePointAt(0).toString(16).toUpperCase().padStart(4, "0"),
-    source: contextSource,
-  };
-}
-
-function parseBlockOptions(source) {
-  const lines = String(source || "").split(/\r?\n/);
-  const options = {
-    lines: [],
-    entries: [],
-    context_directives: [],
-    directive_warnings: [],
-  };
-  let pendingContext = null;
-  for (let sourceLineIndex = 0; sourceLineIndex < lines.length; sourceLineIndex += 1) {
-    const line = lines[sourceLineIndex];
-    const oneBasedLineIndex = sourceLineIndex + 1;
-    const trimmed = line.trim();
-    // Diagnostics are controlled by plugin settings. Strip legacy directives so
-    // they never become learner input and never consume an explicit context turn.
-    if (trimmed === "@diagnostic" || trimmed === "@debug" || trimmed === "@trace") continue;
-
-    const contextDirective = parseContextDirectiveLine(line, oneBasedLineIndex);
-    if (contextDirective) {
-      if (!normalizeSurface(contextDirective.source)) {
-        options.directive_warnings.push({
-          kind: "empty_context_directive",
-          directive_line_index: oneBasedLineIndex,
-          raw_directive: line,
-        });
-        pendingContext = null;
-        continue;
-      }
-      if (pendingContext) {
-        options.directive_warnings.push({
-          kind: "unconsumed_context_directive_replaced",
-          directive_line_index: pendingContext.directive_line_index,
-          replacement_directive_line_index: oneBasedLineIndex,
-          raw_directive: pendingContext.raw_directive,
-        });
-      }
-      pendingContext = contextDirective;
-      options.context_directives.push({ ...contextDirective });
-      continue;
-    }
-
-    options.lines.push(line);
-    const isSourceRow = Boolean(normalizeSurface(line));
-    const appliedContext = pendingContext && isSourceRow
-      ? {
-        ...pendingContext,
-        target_source_line_index: oneBasedLineIndex,
-      }
-      : null;
-    const entry = {
-      line,
-      source_line_index: oneBasedLineIndex,
-      context_directive: appliedContext,
-      context: appliedContext ? { turns: [{ ...appliedContext }] } : null,
-    };
-    options.entries.push(entry);
-    if (isSourceRow) pendingContext = null;
-  }
-  if (pendingContext) {
-    options.directive_warnings.push({
-      kind: "dangling_context_directive",
-      directive_line_index: pendingContext.directive_line_index,
-      raw_directive: pendingContext.raw_directive,
-      source: pendingContext.source,
-    });
-  }
-  return options;
-}
-
 function splitJyutping(jyutping) {
   return String(jyutping || "").trim().split(/\s+/).filter(Boolean);
 }
 
-function renderJyutping(container, jyutping) {
-  const wrapper = createEl("span", "canto-jyutping");
-  for (const syllable of splitJyutping(jyutping)) {
-    const match = syllable.match(/^(.*?)([1-6])$/);
-    const syl = createEl("span", `vjp-syl ${match ? `vjp-tone-${match[2]}` : "vjp-no-tone"}`);
-    if (match) {
-      syl.appendChild(createEl("span", "vjp-base", match[1]));
-      syl.appendChild(createEl("span", "vjp-num", match[2]));
-    } else {
-      syl.appendChild(createEl("span", "vjp-base", syllable));
-    }
-    wrapper.appendChild(syl);
-  }
-  container.appendChild(wrapper);
-}
 
-function renderToken(node, container) {
-  const cls = [`canto-token`, `cs-${safeClass(node.label)}`];
-  if (node.label === "neutral") cls.push("canto-token-neutral", "canto-lexical-unknown");
-  const displaySurface = node.display_surface || node.surface;
-  const span = createEl("span", cls.join(" "));
-  span.setAttribute("tabindex", "0");
-  span.setAttribute("data-canto-kind", "token");
-  span.setAttribute("data-canto-surface", displaySurface);
-  span.setAttribute("data-canto-parser-surface", node.surface);
-  span.setAttribute("data-canto-role", node.label);
-  span.setAttribute("data-canto-syntax", node.syntax);
-  span.setAttribute("data-canto-review", node.review);
-  span.setAttribute("title", learnerUiHoverTitleForToken(node));
-
-  if (node.jyutping) renderJyutping(span, node.jyutping);
-  span.appendChild(createEl("span", "canto-hanzi", displaySurface));
-  container.appendChild(span);
-}
-
-function collectConstructionLayout(node) {
-  const leaves = [];
-  const spans = [];
-
-  function walk(current, depth) {
-    if (!current) return { start: leaves.length, end: leaves.length };
-    if (current.kind !== "construction") {
-      const start = leaves.length;
-      leaves.push(current);
-      return { start, end: leaves.length };
-    }
-
-    const start = leaves.length;
-    for (const child of current.children || []) walk(child, depth + 1);
-    const end = leaves.length;
-    spans.push({ node: current, start, end, depth, width: end - start });
-    return { start, end };
-  }
-
-  walk(node, 0);
-  return { leaves, spans };
-}
-
-function renderConstructionLeaf(node, container) {
-  if (!node) return;
-  if (node.kind === "text") {
-    container.appendChild(document.createTextNode(node.display_text || node.text));
-    return;
-  }
-  if (node.kind === "token") {
-    renderToken(node, container);
-  }
-}
-
-function renderConstruction(node, container) {
-  const { leaves, spans } = collectConstructionLayout(node);
-  const cls = ["canto-pattern", "canto-pattern-layered", `canto-construction-${safeClass(node.type)}`];
-  if (node.primary) cls.push("canto-primary-phrase", `canto-primary-${safeClass(node.primary)}`);
-  const outer = createEl("span", cls.join(" "));
-  outer.setAttribute("data-canto-kind", "construction");
-  outer.setAttribute("data-canto-construction", node.type);
-  outer.setAttribute("title", learnerUiHoverTitleForConstruction(node));
-  outer.style.gridTemplateColumns = leaves.length ? `repeat(${leaves.length}, max-content)` : "max-content";
-
-  leaves.forEach((leaf, index) => {
-    const cell = createEl("span", "canto-pattern-leaf");
-    cell.style.gridColumn = `${index + 1} / ${index + 2}`;
-    cell.style.gridRow = "1";
-    renderConstructionLeaf(leaf, cell);
-    outer.appendChild(cell);
-  });
-
-  spans
-    .filter((span) => span.width > 0)
-    .sort((a, b) => {
-      if (b.width !== a.width) return b.width - a.width;
-      if (a.start !== b.start) return a.start - b.start;
-      return a.depth - b.depth;
-    })
-    .forEach((span, layerIndex) => {
-      const layer = createEl("span", `canto-construction-layer canto-construction-depth-${Math.min(layerIndex, 7)}`);
-      layer.style.gridColumn = `${span.start + 1} / ${span.end + 1}`;
-      layer.style.gridRow = `${layerIndex + 2}`;
-      layer.setAttribute("data-canto-construction", span.node.type);
-      layer.setAttribute("title", learnerUiHoverTitleForConstructionLayer(span.node));
-      layer.appendChild(createEl("span", "canto-construction-line"));
-      layer.appendChild(createEl("span", "canto-pattern-label", span.node.label));
-      outer.appendChild(layer);
-    });
-
-  container.appendChild(outer);
-}
-
-function renderNode(node, container, options = {}) {
-  if (!node) return;
-  if (node.kind === "text") {
-    container.appendChild(document.createTextNode(node.text));
-    return;
-  }
-  if (node.kind === "token") {
-    renderToken(node, container);
-    return;
-  }
-  if (node.kind === "construction") {
-    if (shouldCollapseClauseSequenceForDisplay(node, options) || shouldCollapseGreedyWrapperForDisplay(node, options)) {
-      for (const child of node.children || []) renderNode(child, container, options);
-      return;
-    }
-    renderConstruction(node, container);
-  }
-}
-
-function normalizeHexColor(value) {
-  const raw = String(value || "").trim();
-  const short = raw.match(/^#([0-9a-f]{3})$/i);
-  if (short) {
-    return `#${short[1].split("").map((ch) => ch + ch).join("")}`.toUpperCase();
-  }
-  const long = raw.match(/^#([0-9a-f]{6})$/i);
-  return long ? `#${long[1]}`.toUpperCase() : "";
-}
-
-function getMergedRoleColors(settings) {
-  const userColors = settings && settings.roleColors && typeof settings.roleColors === "object"
-    ? settings.roleColors
-    : {};
-  const merged = { ...DEFAULT_ROLE_COLORS };
-  for (const role of ROLE_COLOR_SETTINGS) {
-    let normalized = normalizeHexColor(userColors[role.key]);
-    merged[role.key] = normalized || role.defaultColor;
-  }
-  return merged;
-}
-
-module.exports = class CantoSpanPlugin extends Plugin {
-  async onload() {
-    await this.loadSettings();
-    this.renderedCantoBlocks = new Map();
-    this.syncGlobalSettingClasses();
-    this.startDiagnosticDetailsObserver();
-
-    this.registerMarkdownCodeBlockProcessor("canto-span", (source, element) => {
-      this.renderedCantoBlocks.set(element, source);
-      this.renderBlock(source, element);
-    });
-
-    this.addSettingTab(new CantoSpanSettingTab(this.app, this));
-  }
-
-  onunload() {
-    if (this.diagnosticDetailsObserver) {
-      this.diagnosticDetailsObserver.disconnect();
-      this.diagnosticDetailsObserver = null;
-    }
-    if (typeof document !== "undefined" && document.body) {
-      document.body.classList.remove(GLOBAL_COLOR_ACCESSIBILITY_CLASS);
-      document.body.classList.remove(GLOBAL_DIAGNOSTICS_CLASS);
-      document.body.classList.remove(GLOBAL_DIAGNOSTICS_EXPANDED_CLASS);
-      this.clearRoleColorVariables();
-    }
-  }
-
-  async loadSettings() {
-    const loaded = await this.loadData();
-    this.settings = Object.assign({}, DEFAULT_SETTINGS, loaded || {});
-    this.settings.roleColors = getMergedRoleColors(this.settings);
-  }
-
-  async saveSettings() {
-    await this.saveData(this.settings);
-    this.syncGlobalSettingClasses();
-  }
-
-  syncGlobalSettingClasses() {
-    if (typeof document === "undefined" || !document.body) return;
-    this.pruneRenderedCantoBlocks();
-    document.body.classList.toggle(
-      GLOBAL_COLOR_ACCESSIBILITY_CLASS,
-      Boolean(this.settings && this.settings.showColorAccessibilityMarkers)
-    );
-    document.body.classList.toggle(
-      GLOBAL_DIAGNOSTICS_CLASS,
-      Boolean(this.settings && this.settings.showDiagnostics)
-    );
-    document.body.classList.toggle(
-      GLOBAL_DIAGNOSTICS_EXPANDED_CLASS,
-      Boolean(this.settings && this.settings.expandDiagnostics)
-    );
-    this.syncRoleColorVariables();
-    this.scheduleRenderedDiagnosticDetailsSync();
-  }
-
-  syncRoleColorVariables() {
-    if (typeof document === "undefined" || !document.body) return;
-    const colors = getMergedRoleColors(this.settings);
-    this.settings.roleColors = colors;
-    for (const role of ROLE_COLOR_SETTINGS) {
-      const color = colors[role.key];
-      for (const cssVar of role.cssVars) {
-        document.body.style.setProperty(`--canto-user-color-${cssVar}`, color);
-      }
-    }
-  }
-
-  clearRoleColorVariables() {
-    if (typeof document === "undefined" || !document.body) return;
-    for (const role of ROLE_COLOR_SETTINGS) {
-      for (const cssVar of role.cssVars) {
-        document.body.style.removeProperty(`--canto-user-color-${cssVar}`);
-      }
-    }
-  }
-
-  pruneRenderedCantoBlocks() {
-    if (!this.renderedCantoBlocks) return 0;
-    let pruned = 0;
-    for (const [element] of Array.from(this.renderedCantoBlocks.entries())) {
-      if (!element || !element.isConnected) {
-        this.renderedCantoBlocks.delete(element);
-        pruned += 1;
-      }
-    }
-    return pruned;
-  }
-
-  getRenderedDiagnosticDetails() {
-    this.pruneRenderedCantoBlocks();
-    const detailsSet = new Set();
-    if (this.renderedCantoBlocks) {
-      for (const element of this.renderedCantoBlocks.keys()) {
-        if (!element || !element.isConnected || typeof element.querySelectorAll !== "function") continue;
-        for (const details of Array.from(element.querySelectorAll(".canto-study-block details.canto-diagnostic"))) {
-          detailsSet.add(details);
-        }
-      }
-    }
-    if (typeof document !== "undefined" && typeof document.querySelectorAll === "function") {
-      for (const details of Array.from(document.querySelectorAll(".canto-study-block details.canto-diagnostic"))) {
-        detailsSet.add(details);
-      }
-    }
-    return Array.from(detailsSet);
-  }
-
-  syncRenderedDiagnosticDetails() {
-    const shouldOpen = Boolean(this.settings && this.settings.expandDiagnostics);
-    for (const details of this.getRenderedDiagnosticDetails()) {
-      if (shouldOpen) {
-        details.open = true;
-        details.setAttribute("open", "");
-        details.setAttribute("data-canto-expanded-by-setting", "true");
-      } else {
-        details.open = false;
-        details.removeAttribute("open");
-        details.setAttribute("data-canto-expanded-by-setting", "false");
-      }
-    }
-  }
-
-  scheduleRenderedDiagnosticDetailsSync() {
-    this.syncRenderedDiagnosticDetails();
-    if (typeof requestAnimationFrame === "function") {
-      requestAnimationFrame(() => this.syncRenderedDiagnosticDetails());
-    }
-    if (typeof window !== "undefined" && typeof window.setTimeout === "function") {
-      window.setTimeout(() => this.syncRenderedDiagnosticDetails(), 0);
-    }
-  }
-
-  queueDiagnosticDetailsSync() {
-    if (this.diagnosticDetailsSyncQueued) return;
-    this.diagnosticDetailsSyncQueued = true;
-    const run = () => {
-      this.diagnosticDetailsSyncQueued = false;
-      this.syncRenderedDiagnosticDetails();
-    };
-    if (typeof requestAnimationFrame === "function") {
-      requestAnimationFrame(run);
-    } else if (typeof window !== "undefined" && typeof window.setTimeout === "function") {
-      window.setTimeout(run, 0);
-    } else {
-      run();
-    }
-  }
-
-  startDiagnosticDetailsObserver() {
-    if (typeof MutationObserver === "undefined" || typeof document === "undefined" || !document.body) return;
-    this.diagnosticDetailsObserver = new MutationObserver((mutations) => {
-      for (const mutation of mutations) {
-        for (const node of Array.from(mutation.addedNodes || [])) {
-          if (!node || node.nodeType !== 1) continue;
-          const element = node;
-          if (element.matches && element.matches("details.canto-diagnostic")) {
-            this.queueDiagnosticDetailsSync();
-            return;
-          }
-          if (element.querySelector && element.querySelector("details.canto-diagnostic")) {
-            this.queueDiagnosticDetailsSync();
-            return;
-          }
-        }
-      }
-    });
-    this.diagnosticDetailsObserver.observe(document.body, { childList: true, subtree: true });
-  }
-
-  refreshRenderedBlocks() {
-    if (!this.renderedCantoBlocks) return;
-    this.pruneRenderedCantoBlocks();
-    for (const [element, source] of Array.from(this.renderedCantoBlocks.entries())) {
-      if (!element || !element.isConnected) continue;
-      this.renderBlock(source, element);
-    }
-  }
-
-  async readActiveNoteMarkdown() {
-    const file = this.app.workspace && this.app.workspace.getActiveFile
-      ? this.app.workspace.getActiveFile()
-      : null;
-    if (!file) return { file: null, markdown: "" };
-    const markdown = await this.app.vault.read(file);
-    return { file, markdown };
-  }
-
-  async formatActiveNoteDiagnostics() {
-    const { markdown } = await this.readActiveNoteMarkdown();
-    return formatDiagnosticMarkdownForNote(markdown);
-  }
-
-  async saveActiveNoteDiagnosticsJson() {
-    const { file, markdown } = await this.readActiveNoteMarkdown();
-    if (!file || !file.path) {
-      new Notice("No active note found for Canto Span diagnostics JSON.");
-      return { summaryPath: "", fullPath: "" };
-    }
-    const bundle = diagnosticExportPayloadsForNote(markdown, file.path);
-    const { summaryPath, fullPath } = bundle.paths;
-    if (!summaryPath || !fullPath) {
-      new Notice("Could not determine Canto Span diagnostics JSON paths.");
-      return { summaryPath: "", fullPath: "" };
-    }
-    for (const outputPath of [summaryPath, fullPath]) {
-      const existing = this.app.vault.getAbstractFileByPath(outputPath);
-      if (existing && existing.children) {
-        new Notice(`Cannot write Canto Span diagnostics JSON: ${outputPath} is a folder.`);
-        return { summaryPath: "", fullPath: "" };
-      }
-    }
-    const writeJson = async (outputPath, payload) => {
-      const json = JSON.stringify(payload, null, 2);
-      const existing = this.app.vault.getAbstractFileByPath(outputPath);
-      if (existing) await this.app.vault.modify(existing, json);
-      else await this.app.vault.create(outputPath, json);
-    };
-    // Write the full artifact first so the summary never points to a missing
-    // companion after a successful two-file save.
-    await writeJson(fullPath, bundle.full);
-    await writeJson(summaryPath, bundle.summary);
-    new Notice(`Saved Canto Span acceptance summary and full diagnostics JSON beside the note.`);
-    return { summaryPath, fullPath };
-  }
-
-  async formatActiveNoteWithDiagnostics() {
-    const { file, markdown } = await this.readActiveNoteMarkdown();
-    return formatNoteMarkdownWithDiagnostics(markdown, file && file.path ? file.path : "");
-  }
-
-  renderBlock(source, element) {
-    element.empty();
-    const options = parseBlockOptions(source);
-    const expandDiagnostics = Boolean(this.settings && this.settings.expandDiagnostics);
-    const block = createEl("div", BLOCK_CLASS);
-    block.setAttribute("data-canto-runtime-version", CANTO_SPAN_RUNTIME_VERSION);
-    const titleRow = createEl("div", "canto-title-row");
-    const title = createEl("div", "canto-title");
-    title.appendChild(createEl("span", "canto-title-normal", "Canto Span"));
-    title.appendChild(createEl("span", "canto-title-diagnostic", `Canto Span — Diagnostic v${CANTO_SPAN_RUNTIME_VERSION}`));
-    titleRow.appendChild(title);
-
-    const analyzedLines = options.entries.map((entry) => ({
-      line: entry.line,
-      source_line_index: entry.source_line_index,
-      context_directive: entry.context_directive || null,
-      analysis: analyzeLine(entry.line, entry.context),
-    }));
-    const diagnosticLines = analyzedLines.filter((entry) => normalizeSurface(entry.line));
-    if (diagnosticLines.length) {
-      renderCopyButton(
-        titleRow,
-        () => this.formatActiveNoteWithDiagnostics(),
-        "Copy note + diagnostics",
-        "Copied the active note with Canto Span diagnostics at the top."
-      );
-      renderCopyButton(
-        titleRow,
-        () => this.formatActiveNoteDiagnostics(),
-        "Copy diagnostics only",
-        "Copied all Canto Span diagnostics in this note."
-      );
-      renderActionButton(
-        titleRow,
-        () => this.saveActiveNoteDiagnosticsJson(),
-        "Save diagnostics JSON"
-      );
-      renderCopyButton(
-        titleRow,
-        () => diagnosticLines.map((entry) => formatDiagnosticMarkdown(entry.analysis)).join("\n\n---\n\n"),
-        diagnosticLines.length === 1 ? "Copy block diagnostic" : "Copy block diagnostics",
-        diagnosticLines.length === 1 ? "Copied this Canto Span block diagnostic." : "Copied this Canto Span block's diagnostics."
-      );
-    }
-
-    const legend = createEl("div", "canto-context-review-boundaries");
-    legend.appendChild(createEl("span", "canto-legend-normal", "Parent construction spans wrap child learner-role tokens. Hover/tap tokens for Jyutping."));
-    legend.appendChild(createEl("span", "canto-legend-diagnostic", "Diagnostic mode is enabled in plugin settings: tokenization, Jyutping audit, generated slots, construction match source, and review traces are shown. Use Copy note + diagnostics to copy the whole active note with generated diagnostics at the top, Copy diagnostics only for the generated batch, Save diagnostics JSON to write both a compact acceptance summary and a companion full-diagnostics file beside the active note, or Copy block diagnostic for one block."));
-    const rendered = createEl("div", "canto-rendered");
-
-    for (const { line, analysis, context_directive: contextDirective } of analyzedLines) {
-      if (contextDirective) {
-        const contextMeta = createEl("div", "canto-context-directive-diagnostic canto-diagnostic-ui-only");
-        contextMeta.setAttribute("data-canto-context-turn-id", contextDirective.id);
-        contextMeta.setAttribute("data-canto-context-directive-line", String(contextDirective.directive_line_index));
-        contextMeta.setAttribute("data-canto-context-target-line", String(contextDirective.target_source_line_index));
-        contextMeta.textContent = `Explicit context → ${contextDirective.source}`;
-        rendered.appendChild(contextMeta);
-      }
-      for (const warning of analysis.warnings) {
-        const warningEl = createEl("div", "canto-context-review-warning", warning);
-        rendered.appendChild(warningEl);
-      }
-      renderParserShadowRepairs(analysis, rendered);
-
-      const lineNormal = createEl("div", "canto-line canto-line-normal");
-      for (const node of analysis.nodes) renderNode(node, lineNormal, { showDiagnostics: false });
-      rendered.appendChild(lineNormal);
-
-      const lineDiagnostic = createEl("div", "canto-line canto-line-diagnostic");
-      for (const node of analysis.nodes) renderNode(node, lineDiagnostic, { showDiagnostics: true });
-      rendered.appendChild(lineDiagnostic);
-
-      if (normalizeSurface(line)) renderDiagnosticDetails(analysis, rendered, expandDiagnostics);
-    }
-
-    block.appendChild(titleRow);
-    block.appendChild(legend);
-    block.appendChild(rendered);
-    element.appendChild(block);
-    this.scheduleRenderedDiagnosticDetailsSync();
-  }
-};
-
-class CantoSpanSettingTab extends PluginSettingTab {
-  constructor(app, plugin) {
-    super(app, plugin);
-    this.plugin = plugin;
-  }
-
-  display() {
-    const { containerEl } = this;
-    containerEl.empty();
-
-    containerEl.createEl("h2", { text: "Canto Span" });
-
-    new Setting(containerEl)
-      .setName("Show parse diagnostics")
-      .setDesc("Show tokenization, Jyutping audit, generated slot candidates, final construction tree, and trace source below each rendered line.")
-      .addToggle((toggle) => toggle
-        .setValue(Boolean(this.plugin.settings.showDiagnostics))
-        .onChange(async (value) => {
-          this.plugin.settings.showDiagnostics = value;
-          await this.plugin.saveSettings();
-          this.display();
-        }));
-
-    new Setting(containerEl)
-      .setName("Expand diagnostics by default")
-      .setDesc("Open diagnostic details automatically when parse diagnostics are shown. This can be set before enabling parse diagnostics.")
-      .addToggle((toggle) => toggle
-        .setValue(Boolean(this.plugin.settings.expandDiagnostics))
-        .onChange(async (value) => {
-          this.plugin.settings.expandDiagnostics = value;
-          await this.plugin.saveSettings();
-          this.display();
-        }));
-
-    new Setting(containerEl)
-      .setName("Show color accessibility markers")
-      .setDesc("Adds non-color learner-role markers for protanopia, deuteranopia, tritanopia, and grayscale support. When off, learner-role styling uses color only.")
-      .addToggle((toggle) => toggle
-        .setValue(Boolean(this.plugin.settings.showColorAccessibilityMarkers))
-        .onChange(async (value) => {
-          this.plugin.settings.showColorAccessibilityMarkers = value;
-          await this.plugin.saveSettings();
-          this.display();
-        }));
-
-    containerEl.createEl("h3", { text: "Role colors" });
-    const colorDesc = containerEl.createEl("p", { cls: "setting-item-description" });
-    colorDesc.textContent = "Choose learner-role colors. Changes apply to already-rendered canto-span blocks immediately. Reset restores the neon default for that role.";
-
-    this.plugin.settings.roleColors = getMergedRoleColors(this.plugin.settings);
-    for (const role of ROLE_COLOR_SETTINGS) {
-      new Setting(containerEl)
-        .setName(role.label)
-        .setDesc(`CSS variable${role.cssVars.length > 1 ? "s" : ""}: ${role.cssVars.map((cssVar) => `--canto-color-${cssVar}`).join(", ")}`)
-        .addColorPicker((picker) => picker
-          .setValue(this.plugin.settings.roleColors[role.key] || role.defaultColor)
-          .onChange(async (value) => {
-            const normalized = normalizeHexColor(value) || role.defaultColor;
-            this.plugin.settings.roleColors[role.key] = normalized;
-            await this.plugin.saveSettings();
-          }))
-        .addButton((button) => button
-          .setButtonText("Reset")
-          .setTooltip(`Reset ${role.label} to ${role.defaultColor}`)
-          .onClick(async () => {
-            this.plugin.settings.roleColors[role.key] = role.defaultColor;
-            await this.plugin.saveSettings();
-            this.display();
-          }));
-    }
-
-    new Setting(containerEl)
-      .setName("Reset all role colors")
-      .setDesc("Restore every learner-role color to the v0.5.4 neon defaults.")
-      .addButton((button) => button
-        .setButtonText("Reset all")
-        .setWarning()
-        .onClick(async () => {
-          this.plugin.settings.roleColors = { ...DEFAULT_ROLE_COLORS };
-          await this.plugin.saveSettings();
-          this.display();
-        }));
-
-    const note = containerEl.createEl("p", { cls: "setting-item-description" });
-    note.textContent = "Review signal: prefer generative_template and slot-based traces. Treat legacy_surface_rule and surface_specific_phrase_rule as places to simplify later.";
-  }
-}
+module.exports = createCantoSpanPlugin({
+  CANTO_SPAN_RUNTIME_VERSION,
+  analyzeLine,
+  normalizeSurface,
+  formatDiagnosticMarkdownForNote,
+  diagnosticExportPayloadsForNote,
+  formatNoteMarkdownWithDiagnostics,
+  formatDiagnosticMarkdown,
+  diagnosticSummary,
+  jyutpingAuditSummary,
+  foldedLexicalRepairDisplayRows,
+  normalizationReviewSuggestionDisplayRows,
+  learnerUiHoverAuditSummary,
+  learnerUiHoverAuditRows,
+  diagnosticTokenRows,
+  diagnosticFinalRows,
+  diagnosticLegend,
+  safeClass,
+  nodeDisplaySurface,
+  learnerUiHoverTitleForToken,
+  learnerUiHoverTitleForConstruction,
+  learnerUiHoverTitleForConstructionLayer,
+  shouldCollapseClauseSequenceForDisplay,
+  shouldCollapseGreedyWrapperForDisplay,
+  parserShadowRepairDisplayRows,
+  parserShadowRepairKindLabel,
+  splitJyutping,
+});
 
 <!-- v0.5.83: Promotes 聽到 / 聽唔到聲 result-potential complements from structural heuristics to named reusable generative templates. -->
 <!-- v0.5.82-r2: Keeps contextual-token structural fixes and corrects 吓 learner hover gloss when active syntax is delimitative_aspect. -->
